requirements.txt
Flask==2.1.1
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dotenv==0.20.0
requests==2.27.1
spotipy==2.19.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
Werkzeug==2.1.1
Flask-Session==0.4.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
requests-oauthlib==1.3.1

